+++ app/Views/maintemp/kompetensi.php
@@ -1,6 +1,6 @@
-<?= $this->extend('maintemp/template');?>
+<?= $this->extend('maintemp/template'); ?>
 
-<?= $this->section('content');?>
+<?= $this->section('content'); ?>
 
 <div class="card card-primary" style="width: auto; margin: 30px;">
     <div class="col-sm-13" style="width: auto; margin: 30px;">
@@ -9,14 +9,14 @@
             <div class="card-body">
 
                 <div class="col">
-                    <?php if(session()->getFlashdata('msg')):?>
-                    <div class="alert alert-success"><?= session()->getFlashdata('msg') ?></div>
-                    <?php endif;?>
+                    <?php if (session()->getFlashdata('msg')) : ?>
+                        <div class="alert alert-success"><?= session()->getFlashdata('msg') ?></div>
+                    <?php endif; ?>
                 </div>
 
                 <div class="col">
                     <div class="row">
-                        <a href="<?php echo base_url();?>/mankomp/tambahkomp" class="btn btn-primary">Tambah
+                        <a href="<?php echo base_url(); ?>/mankomp/tambahkomp" class="btn btn-primary">Tambah
                             Kompetensi</a>
                     </div>
                 </div>
@@ -27,50 +27,46 @@
                     </div>
                 </div>
 
-                <?php if(isset($data_komp)&&($data_komp=="kosong")){
-                    ?>
+                <?php if (isset($data_komp) && ($data_komp == "kosong")) {
+                ?>
 
-                <div class="alert alert-danger">Data kompetensi belum ada. <a
-                        href="<?= base_url();?>/mankomp/tambahkomp">Klik
-                        di sini untuk menambah data kompetensi</a></div>
-                <?php }else{ ?>
+                    <div class="alert alert-danger">Data kompetensi belum ada. <a href="<?= base_url(); ?>/mankomp/tambahkomp">Klik
+                            di sini untuk menambah data kompetensi</a></div>
+                <?php } else { ?>
 
-                <table id="tabledata" class="display table table-bordered table-hover">
-                    <thead>
-                        <tr>
-                            <th>No</th>
-                            <th>Kode</th>
-                            <th>Deskripsi</th>
-                            <th>Kategori</th>
-                            <th>Parent</th>
-                            <th>Aksi</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <?php
-                                    $i=1; 
-                                    foreach ($data_komp as $komp) : 
-                                    ?>
-                        <tr>
-                            <td><?php echo $i;$i++;?></td>
-                            <td><?= $komp['komp_code'];?></td>
-                            <td><?= $komp['komp_desc'];?></td>
-                            <td><?= $komp['komp_cat'];?></td>
-                            <td><?= $komp['komp_parent'];?></td>
-                            <td style="text-align: center"><a
-                                    href="<?php echo base_url();?>/mankomp/ubahkomp/<?=$komp['komp_id'];?>"
-                                    class="btn btn-warning"> <i class="fas fa-file-signature"></i> Ubah</a>
-                                <a href="<?php echo base_url();?>/mankomp/hapuskomp/<?=$komp['komp_id'];?>"
-                                    onclick="return confirm('Apakah anda yakin akan menghapus data kompetensi?')"
-                                    class="btn btn-danger"> <i class="fas fa-trash"></i>
-                                    Hapus</a>
-                            </td>
-                        </tr>
-                        <?php 
-                                    endforeach 
-                                    ?>
-                    </tbody>
-                </table>
+                    <table id="tabledata" class="display table table-bordered table-hover">
+                        <thead>
+                            <tr>
+                                <th>No</th>
+                                <th>Kode</th>
+                                <th>Deskripsi</th>
+                                <th>Kategori</th>
+                                <th>Parent</th>
+                                <th>Aksi</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php
+                            $i = 1;
+                            foreach ($data_komp as $komp) :
+                            ?>
+                                <tr>
+                                    <td><?php echo $i;
+                                        $i++; ?></td>
+                                    <td><?= $komp['komp_code']; ?></td>
+                                    <td><?= $komp['komp_desc']; ?></td>
+                                    <td><?= $komp['komp_cat']; ?></td>
+                                    <td><?= $komp['komp_parent']; ?></td>
+                                    <td style="text-align: center"><a href="<?php echo base_url(); ?>/mankomp/ubahkomp/<?= $komp['komp_id']; ?>" class="btn btn-warning"> <i class="fas fa-file-signature"></i> Ubah</a>
+                                        <a href="<?php echo base_url(); ?>/mankomp/hapuskomp/<?= $komp['komp_id']; ?>" onclick="return confirm('Apakah anda yakin akan menghapus data kompetensi?')" class="btn btn-danger"> <i class="fas fa-trash"></i>
+                                            Hapus</a>
+                                    </td>
+                                </tr>
+                            <?php
+                            endforeach
+                            ?>
+                        </tbody>
+                    </table>
                 <?php } ?>
             </div>
         </div>
@@ -79,4 +75,4 @@
 <!-- /.content-wrapper -->
 <!-- Control Sidebar -->
 
-<?= $this->endSection();?>
+<?= $this->endSection(); ?>
