+++ app/Views/maintemp/manjadwalsidang.php
@@ -1,6 +1,6 @@
-<?= $this->extend('maintemp/template');?>
+<?= $this->extend('maintemp/template'); ?>
 
-<?= $this->section('content');?>
+<?= $this->section('content'); ?>
 
 <div class="card card-primary" style="width: auto; margin: 30px;">
     <div class="col-sm-13" style="width: auto; margin: 30px;">
@@ -9,14 +9,14 @@
             <div class="card-body">
 
                 <div class="col">
-                    <?php if(session()->getFlashdata('msg')):?>
-                    <div class="alert alert-success"><?= session()->getFlashdata('msg') ?></div>
-                    <?php endif;?>
+                    <?php if (session()->getFlashdata('msg')) : ?>
+                        <div class="alert alert-success"><?= session()->getFlashdata('msg') ?></div>
+                    <?php endif; ?>
                 </div>
 
                 <div class="col">
                     <div class="row">
-                        <a href="<?php echo base_url();?>/mantugasakhir" class="btn btn-primary">Kembali ke daftar
+                        <a href="<?php echo base_url(); ?>/mantugasakhir" class="btn btn-primary">Kembali ke daftar
                             praktek keinsinyuran.</a>
                     </div>
                 </div>
@@ -27,44 +27,42 @@
                     </div>
                 </div>
 
-                <?php if(isset($data_js)&&($data_js=="kosong")){
-                    ?>
+                <?php if (isset($data_js) && ($data_js == "kosong")) {
+                ?>
 
-                <div class="alert alert-danger">Data jadwal sidang belum ada. <a
-                        href="<?= base_url();?>/mantugasakhir/tambahjadwal/<?= $ta_id;?>/<?= $user_id;?>">Klik
-                        di sini untuk menambah jadwal sidang</a></div>
-                <?php }else{ ?>
+                    <div class="alert alert-danger">Data jadwal sidang belum ada. <a href="<?= base_url(); ?>/mantugasakhir/tambahjadwal/<?= $ta_id; ?>/<?= $user_id; ?>">Klik
+                            di sini untuk menambah jadwal sidang</a></div>
+                <?php } else { ?>
 
-                <table id="tabledata" class="display table table-bordered table-hover">
-                    <thead>
-                        <tr>
-                            <th>No</th>
-                            <th>Ruang Sidang</th>
-                            <th>Tanggal Sidang</th>
-                            <th>Aksi</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <?php
-                                    $i=1; 
-                                    foreach ($data_js as $ta) : 
-                                    ?>
-                        <tr>
-                            <td><?php echo $i;$i++;?></td>
-                            <td><?= $ta['sidang_ruang'];?></td>
-                            <td><?= $ta['sidang_tanggal'];?></td>
-                            <td style="text-align: center">
-                                <a href="<?php echo base_url();?>/mantugasakhir/hapusjadwal/<?=$ta['sidang_id'];?>/<?= $ta_id;?>/<?=$ta['user_id'];?>"
-                                    onclick="return confirm('Apakah anda yakin akan menghapus jadwal sidang?')"
-                                    class="btn btn-danger"> <i class="fas fa-trash"></i>
-                                    Hapus</a>
-                            </td>
-                        </tr>
-                        <?php 
-                                    endforeach 
-                                    ?>
-                    </tbody>
-                </table>
+                    <table id="tabledata" class="display table table-bordered table-hover">
+                        <thead>
+                            <tr>
+                                <th>No</th>
+                                <th>Ruang Sidang</th>
+                                <th>Tanggal Sidang</th>
+                                <th>Aksi</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php
+                            $i = 1;
+                            foreach ($data_js as $ta) :
+                            ?>
+                                <tr>
+                                    <td><?php echo $i;
+                                        $i++; ?></td>
+                                    <td><?= $ta['sidang_ruang']; ?></td>
+                                    <td><?= $ta['sidang_tanggal']; ?></td>
+                                    <td style="text-align: center">
+                                        <a href="<?php echo base_url(); ?>/mantugasakhir/hapusjadwal/<?= $ta['sidang_id']; ?>/<?= $ta_id; ?>/<?= $ta['user_id']; ?>" onclick="return confirm('Apakah anda yakin akan menghapus jadwal sidang?')" class="btn btn-danger"> <i class="fas fa-trash"></i>
+                                            Hapus</a>
+                                    </td>
+                                </tr>
+                            <?php
+                            endforeach
+                            ?>
+                        </tbody>
+                    </table>
                 <?php } ?>
             </div>
         </div>
@@ -73,4 +71,4 @@
 <!-- /.content-wrapper -->
 <!-- Control Sidebar -->
 
-<?= $this->endSection();?>
+<?= $this->endSection(); ?>
