+++ app/Views/maintemp/manpenguji.php
@@ -1,6 +1,6 @@
-<?= $this->extend('maintemp/template');?>
+<?= $this->extend('maintemp/template'); ?>
 
-<?= $this->section('content');?>
+<?= $this->section('content'); ?>
 
 <div class="card card-primary" style="width: auto; margin: 30px;">
     <div class="col-sm-13" style="width: auto; margin: 30px;">
@@ -9,14 +9,14 @@
             <div class="card-body">
 
                 <div class="col">
-                    <?php if(session()->getFlashdata('msg')):?>
-                    <div class="alert alert-success"><?= session()->getFlashdata('msg') ?></div>
-                    <?php endif;?>
+                    <?php if (session()->getFlashdata('msg')) : ?>
+                        <div class="alert alert-success"><?= session()->getFlashdata('msg') ?></div>
+                    <?php endif; ?>
                 </div>
 
                 <div class="col">
                     <div class="row">
-                        <a href="<?php echo base_url();?>/mantugasakhir" class="btn btn-primary">Kembali ke Daftar
+                        <a href="<?php echo base_url(); ?>/mantugasakhir" class="btn btn-primary">Kembali ke Daftar
                             Praktek Keinsinyuran</a>
                     </div>
                 </div>
@@ -29,8 +29,7 @@
 
                 <div class="col">
                     <div class="row">
-                        <a href="<?php echo base_url();?>/mantugasakhir/tambahpenguji/<?= $ta_id;?>/<?= $user_id;?>"
-                            class="btn btn-primary">Atur Penguji</a>
+                        <a href="<?php echo base_url(); ?>/mantugasakhir/tambahpenguji/<?= $ta_id; ?>/<?= $user_id; ?>" class="btn btn-primary">Atur Penguji</a>
                     </div>
                 </div>
 
@@ -40,42 +39,40 @@
                     </div>
                 </div>
 
-                <?php if(isset($data_uji)&&($data_uji=="kosong")){
-                    ?>
+                <?php if (isset($data_uji) && ($data_uji == "kosong")) {
+                ?>
 
-                <div class="alert alert-danger">Data penguji proyek akhir belum ada. <a
-                        href="<?= base_url();?>/mantugasakhir/tambahpenguji/<?= $ta_id;?>/<?= $user_id;?>">Klik
-                        di sini untuk mengatur data penguji</a></div>
-                <?php }else{ ?>
+                    <div class="alert alert-danger">Data penguji proyek akhir belum ada. <a href="<?= base_url(); ?>/mantugasakhir/tambahpenguji/<?= $ta_id; ?>/<?= $user_id; ?>">Klik
+                            di sini untuk mengatur data penguji</a></div>
+                <?php } else { ?>
 
-                <table id="tabledata" class="display table table-bordered table-hover">
-                    <thead>
-                        <tr>
-                            <th>No</th>
-                            <th>Nama Penguji</th>
-                            <th>Aksi</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        <?php
-                                    $i=1; 
-                                    foreach ($data_uji as $ta) : 
-                                    ?>
-                        <tr>
-                            <td><?php echo $i;$i++;?></td>
-                            <td><?= !empty($ta['FullName']) ? $ta['FullName'] : "Belum ada penguji." ;?></td>
-                            <td style="text-align: center">
-                                <a href="<?php echo base_url();?>/mantugasakhir/hapusuji/<?=$ta['ta_id'];?>/<?= $ta_id;?>/<?=$ta['user_id'];?>"
-                                    onclick="return confirm('Apakah anda yakin akan menghapus data penguji?')"
-                                    class="btn btn-danger"> <i class="fas fa-trash"></i>
-                                    Hapus</a>
-                            </td>
-                        </tr>
-                        <?php 
-                                    endforeach 
-                                    ?>
-                    </tbody>
-                </table>
+                    <table id="tabledata" class="display table table-bordered table-hover">
+                        <thead>
+                            <tr>
+                                <th>No</th>
+                                <th>Nama Penguji</th>
+                                <th>Aksi</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php
+                            $i = 1;
+                            foreach ($data_uji as $ta) :
+                            ?>
+                                <tr>
+                                    <td><?php echo $i;
+                                        $i++; ?></td>
+                                    <td><?= !empty($ta['FullName']) ? $ta['FullName'] : "Belum ada penguji."; ?></td>
+                                    <td style="text-align: center">
+                                        <a href="<?php echo base_url(); ?>/mantugasakhir/hapusuji/<?= $ta['ta_id']; ?>/<?= $ta_id; ?>/<?= $ta['user_id']; ?>" onclick="return confirm('Apakah anda yakin akan menghapus data penguji?')" class="btn btn-danger"> <i class="fas fa-trash"></i>
+                                            Hapus</a>
+                                    </td>
+                                </tr>
+                            <?php
+                            endforeach
+                            ?>
+                        </tbody>
+                    </table>
                 <?php } ?>
             </div>
         </div>
@@ -84,4 +81,4 @@
 <!-- /.content-wrapper -->
 <!-- Control Sidebar -->
 
-<?= $this->endSection();?>
+<?= $this->endSection(); ?>
